+++ massiv-test/tests/Test/Massiv/Array/Delayed/StreamSpec.hs
@@ -1,0 +1,8 @@
+module Test.Massiv.Array.Delayed.StreamSpec (spec) where
+
+import Test.Massiv.Core
+import Test.Massiv.Array.Delayed
+
+
+spec :: Spec
+spec = delayedStreamSpec