+++ massiv/CHANGELOG.md
@@ -4,6 +4,10 @@
   * `filterS`, `filterM`, `ifilterS`, `ifilterM`
   * `mapMaybeS`, `mapMaybeM`, `imapMaybeS`, `imapMaybeM`
   * `unfoldr`, `unfoldrN`
+  * `takeS` and `dropS`
+* Deprecated `traverseAR`, `itraverseAR`, `traversePrimR`, `itraversePrimR` (not feasible
+  to keep duplicate functions just for representation, `TypeApplications` or
+  `ScopedVariables` should be used instead.)
 
 * Fix performance issue with copying of unboxed arrays and initialization of storable array.
 * Addition of `unsafeLoadIntoS`, `unsafeLoadInto` and `maxSize`